+++ NEWS.md
@@ -1,3 +1,5 @@
+# themis (development version)
+
 # themis 0.1.3
 
 * Steps that use nearest neighbors gives cleaner errors.