+++ NEWS.md
@@ -1,4 +1,4 @@
-# themis (development version)
+# themis 0.1.4
 
 * export `smote()`, `adasyn()`, and `bsmote()` functions to users.